requirements.txt
# local package
-e .

# external requirements
boto3
modelstore
scikit-learn
pandas
matplotlib
click
Sphinx
coverage
flake8
python-dotenv>=0.5.1
